requirements.txt
PyYAML~=6.0.1
python-dotenv~=1.0.1
pathlib
typing